+++ CHANGELOG.md
@@ -7,3 +7,14 @@
 - Popup interface
 - Toggle switch logic
 - TypeScript configuration
+
+## [0.2.0] - 12/08/2026
+
+### Added
+
+- Current active tab detection
+- Cookie retrieval and removal using the Chrome Cookies API
+
+### Changed
+
+- Toggle switch now triggers cookie removal when enabled